backend/requirements.txt
six==1.10.0
tornado==4.5.1
websocket-client==0.44.0
